requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
clickhouse-driver==0.2.6
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
slack-sdk==3.21.3
psutil==5.9.5
prometheus-client==0.17.1
